backend/pyproject.toml
[project]
name = "focusflow-brain"
version = "0.1.0"
description = "Intelligent backend for FocusFlow"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "websockets",
    "python-dotenv",
    "groq",
    "pydantic",
    "pydantic-settings",
    "langgraph>=1.1.1",
    "langchain-groq>=1.1.2",
    "dspy-ai>=3.1.3",
]

[tool.uv]
managed = true
